backend.laso/requirements.txt
aiosqlite==0.22.1
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
billiard==4.2.4
black==24.1.1
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.1
cryptography==42.0.2
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
et_xmlfile==2.0.0
Faker==22.0.0
fastapi==0.109.0
flake8==7.0.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.8.0
mypy_extensions==1.1.0
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.1.4
passlib==1.7.4
pathspec==1.0.3
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.2.0
PyJWT==2.8.0
pypdf==4.0.1
pypng==0.20220715.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.3
qrcode==7.4.2
redis==5.0.1
reportlab==4.0.9
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.27.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
XlsxWriter==3.1.9
